pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "electronic-bot"
version = "0.0.1"
requires-python = ">=3.10.4"
dependencies = [
        "openai",
        "python-dotenv==1.0.0",
        "tiktoken==0.5.1",
        "panel==1.3.1",
        "langchain",
        "chromadb==0.4.3",
        "pypdf",
        "langchain[docarray]"
]